chatbot-backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "openai-agents>=0.7.0",
    "openai-chatkit>=1.6.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.32",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.11",
]
